Move misc/task_queue.py to db/mongodb
